Disable satisfied-skip-solve on 32-bit linux
The default install of Conda on 32 bit linux uses repo.anaconda.com/miniconda; those 32 bit builds are very dated and so do not support the --satisfied-skip-solve flag

The latest 32 bit linux build (https://repo.anaconda.com/miniconda/Miniconda3-latest-Linux-x86.sh) is 4.5.12, which pre-dates the introduction of --satisfied-skip-solve, which was introduced in 4.6.0 (https://docs.conda.io/projects/conda/en/latest/release-notes.html\#version-4.6.0).

This is a somewhat draconian approach, as someone could have a newer build of Conda on a 32 bit linux system, and this change prevents them from using --satisfied-skip-solve.

Even so, I think this change is worthwhile because (a) the current state means a default install of PyCall (which sets this flag) on 32 bit linux fails (b) preventing --satisfied-skip-solve does not change correctness, just performance (c) there are probably not a lot of people hitting this code path.